server/requirements.txt
asgiref==3.6.0
astroid==2.14.1
autopep8==2.0.1
dill==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.3.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mongoengine==0.26.0
pep8==1.7.1
platformdirs==3.0.0
pycodestyle==2.10.0
pykerberos==1.2.4
pylint==2.16.1
pymongo==4.3.3
python-snappy==0.6.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
